// src/app/Models/HoliDayBalanceFlow.ts
import { DayHourMinuteClass } from "./DayHourMinuteClass";

// Generated by https://quicktype.io

export interface HoliDayBalanceFlow {
    DateB: string;
    DateE: string;
    Max: number;
    Use: number;
    UseDayHourMinute: DayHourMinuteClass;
    FlowUse: number;
    FlowUseDayHourMinute:DayHourMinuteClass;
    Balance: number;
    BalanceDayHourMinute:DayHourMinuteClass;
    BalanceReal: number;
    BalanceRealDayHourMinute:DayHourMinuteClass;
    HoliDayAddition: HoliDay[];
    HoliDaySubtract: HoliDay[];
    AbsAddition: AbsAdditionClass[];
    AbsDeduction: AbsDeduction[];
    ExceptionUse: number;
    AbsFlowApps: any[];
    HoliDayKindID: number;
    HoliDayKindCode: string;
    HoliDayKindNameC: string;
    Sort: number;
    Unit: string;
    HoliDay: HoliDay[];
}

export interface HoliDay {
    HoliDayID: number;
    HoliDayCode: string;
    HoliDayNameC: string;
    HoliDayUnit: string;
    HoliDayUnitName: string;
    MinNum: number;
    IncludeHoliDay: boolean;
    MaxNum: number;
    CheckRestHour: boolean;
    Sex: string;
    IsDisplay: boolean;
    AbsentUnit: number;
    HoliDayKindID: number;
    HoliDayFlag: string;
    Seq: number;
    AutoCreate: boolean;
    AutoFlowStart: boolean;
    AbsentLeave: boolean;
}



export interface AbsDeduction {
    HoliDayCode: string
    HoliDayNameC: string
    HoliDayUnit: string
    Use: DayHourMinuteClass
    Balance: DayHourMinuteClass
    FlowUse: DayHourMinuteClass
}




// Generated by https://quicktype.io

export interface AbsAdditionClass {
    AbsPlusID:   number;
    EmpID:       string;
    DateB:       string;
    DateE:       string;
    BTime:       string;
    ETime:       string;
    HoliDayID:   number;
    TotalAmount: number;
    AbsAmount:   number;
    RestAmount:  number;
    RestAmountDayHourMinute:  DayHourMinuteClass;
    Note:        string;
    Serno:       string;
    AbsPlusType: any;
    KeyName:     string;
    EventDate:   any;
    HoliDay:     HoliDay;
    AbsTrans:    AbsTran[];
    Year:        number;
}

export interface AbsTran {
    AbsTransID: number;
    AbsPlusID:  string;
    AbsMinusID: null;
    DateB:      string;
    Use:        number;
    Abs:        any[];
}

export interface HoliDay {
    HoliDayID:        number;
    HoliDayCode:      string;
    HoliDayNameC:     string;
    HoliDayUnit:      string;
    HoliDayUnitName:  string;
    MinNum:           number;
    IncludeHoliDay:   boolean;
    MaxNum:           number;
    CheckRestHour:    boolean;
    Sex:              string;
    IsDisplay:        boolean;
    AbsentUnit:       number;
    HoliDayKindID:    number;
    HoliDayFlag:      string;
    Seq:              number;
    AutoCreate:       boolean;
    AutoFlowStart:    boolean;
    AbsentLeave:      boolean;
    Memo:             any;
    IsNotRefRote:     boolean;
    IsDisMinus:       boolean;
    IsDisBalance:     boolean;
    IsDisWithHolding: boolean;
    IsThisYear:       boolean;
}